pyproject.toml
[project]
name = "ukan-weeder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "addict>=2.4.0",
    "albumentations>=2.0.5",
    "dataclasses>=0.8",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "medpy>=0.5.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "perceptual>=0.1",
    "pillow>=11.1.0",
    "ptflops>=0.7.4",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "tensorboardx>=2.6.2.2",
    "tifffile>=2025.3.30",
    "timm>=1.0.15",
    "torch-tb-profiler>=0.4.3",
    "torchmetrics>=1.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.51.3",
    "yapf>=0.43.0",
]
